Extract DB connection into separate file
